refactor(agent-task): extract cook-workspace restore from lifecycle_ops (#9927)
